#71 wip refactoring
